Handle "Method Not Allowed"
According to RFC2616 section 10.4.6:

"The method specified in the Request-Line is not allowed for the resource
identified by the Request-URI. The response MUST include an Allow header
containing a list of valid methods for the requested resource."

Pat knows what methods are allowed given a pattern. For convience, pat
will add the Allow header for requests that match a pattern for a method
other than the method requested and set the Status to "405 Method Not
Allowed".